Moved deploy button
